Add CairoBackgroundService for new hosted services
