+++ Cairo Desktop/Cairo Desktop/CairoApplication.xaml.cs
@@ -1,7 +1,6 @@
 ﻿using CairoDesktop.Application.Interfaces;
 using CairoDesktop.Common;
 using CairoDesktop.Configuration;
-using CairoDesktop.SupportingClasses;
 using ManagedShell.Common.Helpers;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
