+++ Cairo Desktop/CairoDesktop.Infrastructure/DependencyInjection/DependencyInjectionExtensions.cs
@@ -15,6 +15,7 @@ namespace CairoDesktop.Infrastructure.DependencyInjection
         {
             services.AddSingleton<IExtensionService, ExtensionService>();
             services.AddSingleton<IApplicationUpdateService, WinSparkleApplicationUpdateService>();
+            services.AddSingleton<ShellManagerService>();
 
             return services;
         }
